Move ToGlog type to package x.
