+++ conn/node.go
@@ -21,7 +21,6 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"log"
 	"math/rand"
 	"strings"
 	"sync"
@@ -79,24 +78,6 @@ type Node struct {
 	heartbeatsIn  int64
 }
 
-// ToGlog is a logger that forwards the output to glog.
-// TODO(martinmr): move this to a more appropriate package.
-type ToGlog struct {
-}
-
-func (rl *ToGlog) Debug(v ...interface{})                   { glog.V(3).Info(v...) }
-func (rl *ToGlog) Debugf(format string, v ...interface{})   { glog.V(3).Infof(format, v...) }
-func (rl *ToGlog) Error(v ...interface{})                   { glog.Error(v...) }
-func (rl *ToGlog) Errorf(format string, v ...interface{})   { glog.Errorf(format, v...) }
-func (rl *ToGlog) Info(v ...interface{})                    { glog.Info(v...) }
-func (rl *ToGlog) Infof(format string, v ...interface{})    { glog.Infof(format, v...) }
-func (rl *ToGlog) Warning(v ...interface{})                 { glog.Warning(v...) }
-func (rl *ToGlog) Warningf(format string, v ...interface{}) { glog.Warningf(format, v...) }
-func (rl *ToGlog) Fatal(v ...interface{})                   { glog.Fatal(v...) }
-func (rl *ToGlog) Fatalf(format string, v ...interface{})   { glog.Fatalf(format, v...) }
-func (rl *ToGlog) Panic(v ...interface{})                   { log.Panic(v...) }
-func (rl *ToGlog) Panicf(format string, v ...interface{})   { log.Panicf(format, v...) }
-
 // NewNode returns a new Node instance.
 func NewNode(rc *pb.RaftContext, store *raftwal.DiskStorage) *Node {
 	snap, err := store.Snapshot()
@@ -144,7 +125,7 @@ func NewNode(rc *pb.RaftContext, store *raftwal.DiskStorage) *Node {
 			// snapshot.
 			Applied: snap.Metadata.Index,
 
-			Logger: &ToGlog{},
+			Logger: &x.ToGlog{},
 		},
 		// processConfChange etc are not throttled so some extra delta, so that we don't
 		// block tick when applyCh is full
